# adaptation-server/openapi.yaml
openapi: 3.0.3
info:
  title: Adaptation Admin API
  description: API for managing employee adaptation process
  version: 1.0.0
servers:
  - url: /api
    description: Default server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
    Employee:
      type: object
      required:
        - id
        - fullName
        - position
        - department
        - email
        - hireDate
        - adaptationStatus
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        position:
          type: string
        department:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        hireDate:
          type: string
          format: date
        adaptationStatus:
          type: string
          enum: [not_started, in_progress, completed]
        mentorId:
          type: string
          format: uuid
        assignedTrackId:
          type: string
          format: uuid
        startDate:
          type: string
          format: date
        stepProgress:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/StepProgress'
        accessLink:
          type: string
    StepProgress:
      type: object
      properties:
        completed:
          type: boolean
        answers:
          type: object
          additionalProperties:
            type: string
...
    AccessLinkResponse:
      type: object
      required:
        - accessLink
      properties:
        accessLink:
          type: string
security:
  - bearerAuth: []
paths:
...
  /employees:
    get:
      summary: Get all employees
      responses:
        '200':
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
    post:
      summary: Create a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fullName
                - position
                - department
                - email
                - hireDate
              properties:
                fullName:
                  type: string
                position:
                  type: string
                department:
                  type: string
                email:
                  type: string
                  format: email
                phone:
                  type: string
                hireDate:
                  type: string
                  format: date
      responses:
        '201':
          description: Employee created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /employees/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get employee by ID
      responses:
        '200':
          description: Employee details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        '200':
          description: Employee updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete employee
      responses:
        '204':
          description: Employee deleted
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /employees/{id}/mentor:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Assign mentor to employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - mentorId
              properties:
                mentorId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Mentor assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '404':
          description: Employee or mentor not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Remove mentor from employee
      responses:
        '200':
          description: Mentor removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /employees/{id}/track:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Assign track to employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - trackId
                - startDate
              properties:
                trackId:
                  type: string
                  format: uuid
                startDate:
                  type: string
                  format: date
      responses:
        '200':
          description: Track assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '404':
          description: Employee or track not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Remove track from employee
      responses:
        '200':
          description: Track removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /employees/{id}/access-link:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Generate access link for employee
      responses:
        '200':
          description: Access link generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessLinkResponse'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...